requirements.txt
bs4
requests
pdfplumber
lxml
pyyaml
